feat Question Save DTO, Question Converter 개발 Question과 Category 연관관계 맵핑부분이 힘들었음

+++ src/main/kotlin/org/ccondaeapi/infrastructure/repository/CategoryRepository.kt
@@ -3,5 +3,5 @@ package org.ccondaeapi.infrastructure.repository
 import org.ccondaeapi.entity.Category
 import org.springframework.data.jpa.repository.JpaRepository
 
-interface CategoryRepository : JpaRepository<Category, Int> {
+interface CategoryRepository : JpaRepository<Category, Long> {
 }
